+++ Changelog.md
@@ -7,6 +7,8 @@
 - Fix: Password-protected Furnace is not shown as a recipe catalyst for furnace recipes in JEI
 - Fix: Changing the owner of the Inventory Scanner or doors does not change the owner of the other block/half
 - Fix: Projecting disguisable blocks (those that support the disguise module) does not work
+- Fix: Crash when trying to place a door one block below to the build height
+- Fix: Door item gets removed when trying to place a door outside the build height
 - Misc.: More texture changes
 
 --------------------------Changelog for v1.9.0.2-beta1 of SecurityCraft--------------------------